Argh, factor out lambda_response.py.
